Make only intended charts draggable
On a page with other charts, this plugin seemed to reduce the chart padding with the `chart.redraw()` line, and made columns have a dotted border if the values were near 0 (which should happen on draggable charts).  I think if you want to have a draggable chart, you shouldn't have to have these events and styles affect other charts.